fix: rename test file and class name
Renamed every test class so it links correctly to the file tested.
Moved MapKtTest.kt to the correct package
