+++ app/src/androidTest/java/com/github/lookupgroup27/lookup/ui/authentication/SignInKtTest.kt
@@ -20,7 +20,7 @@ import org.junit.Test
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
-class LoginTest : TestCase() {
+class SignInKtTest : TestCase() {
 
   @get:Rule val composeTestRule = createComposeRule()
