+++ app/src/androidTest/java/com/github/lookupgroup27/lookup/ui/overview/LandingKtTest.kt
@@ -17,7 +17,7 @@ import org.junit.Test
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
-class LandingScreenTest {
+class LandingKtTest {
 
   @get:Rule val composeTestRule = createAndroidComposeRule<MainActivity>()
