+++ app/src/androidTest/java/com/github/lookupgroup27/lookup/ui/overview/QuizKtTest.kt
@@ -1,4 +1,4 @@
-package com.github.lookupgroup27.lookup.ui.skytracker
+package com.github.lookupgroup27.lookup.ui.overview
 
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createComposeRule
@@ -11,26 +11,26 @@ import org.junit.Test
 import org.mockito.kotlin.mock
 import org.mockito.kotlin.verify
 
-class SkyTrackerScreenTest {
+class QuizKtTest {
 
   @get:Rule val composeTestRule = createComposeRule()
 
   private val mockNavigationActions: NavigationActions = mock()
 
   @Test
-  fun skyTrackerScreen_displaysSkyTrackerText() {
-    composeTestRule.setContent { SkyTrackerScreen(navigationActions = mockNavigationActions) }
+  fun quizScreen_displaysQuizText() {
+    composeTestRule.setContent { QuizScreen(navigationActions = mockNavigationActions) }
 
-    // Verify that the "Sky Tracker Screen" text is displayed
-    composeTestRule.onNodeWithText("Sky Tracker Screen").assertIsDisplayed()
+    // Verify that the "Quiz Screen" text is displayed
+    composeTestRule.onNodeWithText("Quiz Screen").assertIsDisplayed()
   }
 
   @Test
-  fun skyTrackerScreen_clickBackButton_navigatesBack() {
-    composeTestRule.setContent { SkyTrackerScreen(navigationActions = mockNavigationActions) }
+  fun quizScreen_clickBackButton_navigatesBack() {
+    composeTestRule.setContent { QuizScreen(navigationActions = mockNavigationActions) }
 
     // Perform click on the back button
-    composeTestRule.onNodeWithTag("go_back_button_skyTracker").performClick()
+    composeTestRule.onNodeWithTag("go_back_button_quiz").performClick()
 
     // Verify navigation back action is triggered
     verify(mockNavigationActions).goBack()
